requirements.txt
streamlit
python-docx
lxml
pdfplumber
pypdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
